backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pydantic
python-multipart
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator
